+++ packages/create-lila-app/files/vue-app.eslintrc.vue.js
@@ -1,0 +1,27 @@
+module.exports = {
+  root: true,
+  parser: 'vue-eslint-parser',
+  extends: [
+    'plugin:vue/recommended',
+    'airbnb-base',
+    'plugin:flowtype/recommended',
+    'prettier',
+    'prettier/flowtype',
+  ],
+  plugins: ['vue', 'import', 'node', 'prettier', 'flowtype'],
+  env: {
+    node: true,
+    es6: true,
+    browser: true,
+  },
+  parserOptions: {
+    ecmaVersion: 2017,
+  },
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.vue'],
+      },
+    },
+  },
+};