+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.2 [2019-03-30]
 
+- Add `create-lila-app` plugin.
 - Add `lila-webpack-config` plugin.
 
 ## 0.7.1 [2019-03-11]